feat(layout): 修改默认布局模式为单栏
- 将插件的默认布局模式从双栏改为单栏，以提升新用户的使用体验
- 更新相关逻辑和测试，确保重置功能正确返回单栏布局
- 优化布局组件的样式，保持一致性和响应性
- 添加文档说明，详细记录布局修改的原因和影响

+++ src/background.ts
@@ -150,7 +150,7 @@ chrome.runtime.onInstalled.addListener(async (details) => {
       confirmBeforeDelete: true,
       allowDuplicateTabs: false,
       syncEnabled: true,
-      layoutMode: 'double',
+      layoutMode: 'single',
       showNotifications: false, // 默认关闭通知
       syncStrategy: 'newest',
       deleteStrategy: 'everywhere',
